:bug: Improve the consistency of the behaviour of `copy`  (#201)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Apache-2.0
-->
### Description

In some cases such as the copy of files to destination which do not
exist, the behaviour of `filesystem.Copy()` as changed to better aligned
with `cp -r` was not particularly straightforward and potentially
misleading

e.g.
filesystem.Copy("source.txt","does-not-exist") and
filesystem.Copy("source.txt",".does-not-exist") had different behaviours
since the type of destination was interpreted differently.

The intention of this PR is to clarify the behaviour of Copy in those
circumstances and ensure the behaviour is more intuitive.


### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [x]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).