:sparkles: [`http`] Add support for using a custom HTTP client as the base for the retryable HTTP client (#386)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Apache-2.0
-->
### Description

<!--
Please add any detail or context that would be useful to a reviewer.
-->

Add support for using a custom HTTP client as the base for the retryable
HTTP client.

This allows you to have retryable clients with a different underlying
client such as an client with OAuth headers.

There is a wrapper for the common case of adding OAuth headers as well
as functions for passing any HTTP client as the base for the retryable
client.

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [X]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).