:bug: Make sure that loggers added in NewMultipleLoggers don't fail with 'No Logger Source' (#451)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Apache-2.0
-->
### Description

<!--
Please add any detail or context that would be useful to a reviewer.
-->

Make sure that loggers added in NewMultipleLoggers don't fail with 'No
Logger Source'.

NewMultipleLoggers calls Append to add the loggers, this tries to call
c.setLoggerSource but at this point the base logger hasn't had it's
source set. This change makes sure that the source is set earlier in the
process.

It also adds a test that initialises the loggers with NewMultipleLoggers
as there wasn't a test which is why this problem was missed.

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [x]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).