:bug: Make sure that configuration defaults are not over-ridden with blank values when reading in a config file that doesn't specify fields (#506)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Apache-2.0
-->
### Description

<!--
Please add any detail or context that would be useful to a reviewer.
-->

If using a config file then it was overriding the default values that
had already been added to the session. This is fine except if you don't
specify a value in the config because it is replaced with an empty value
and the defaults are not kept.

With the newly added test, if you do it the old way the test fails.

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [x]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).