:sparkles: `encryption` Add support for encrypting content using hybrid RSA/AES encryption (#562)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Apache-2.0
-->
### Description

<!--
Please add any detail or context that would be useful to a reviewer.
-->

This change provides a method of encrypting a payload using hybrid RSA
AES encryption. The payload itself is encrypted with the AES key which
is fast but symmetric meaning both parties require the key. Therefore
the key is encrypted with asymmetric RSA encryption and the encrypted
key is included in the payload itself. RSA is not efficient on larger
payloads so this method allows for fast encryption whilst also being
asymmetric.

Information on hybrid RSA/AES encryption can be found here
https://www.ijrar.org/papers/IJRAR23B1852.pdf

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [x]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).

---------

Co-authored-by: Adrien CABARBAYE <adrien.cabarbaye@arm.com>