:bug: `[filesystem]` fix `bad file descriptor error` on Linux when using Touch (#570)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Apache-2.0
-->
### Description

- On linux the Touch should set file's permissions to `664`. Currently
permissions are set to `600` leading to some bad file descriptor errors
when reading or writing to the file.



### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [x]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).