:sparkles: `safeio` Add support for cancelling readers that make blocking kernel reads during copying (#703)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Apache-2.0
-->
### Description

<!--
Please add any detail or context that would be useful to a reviewer.
-->

Add support for cancelling readers that make blocking kernel reads
during copying to safeio submodule.

When copying, a lot of the time things were in the kernel read and
therefore since the cancel in the other contextual readers only cancel
before a read, they weren't be cancelled.

This allows readclosers to be used and makes it so that context
cancellation will call their close functions which will stop the kernel
read.

https://man7.org/linux/man-pages/man2/read.2.html

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [x]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).