:bug: `parallelisation` Ensure that execution options are propagated to compound execution group when creating a PriorityExecutionGroup (#745)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Apache-2.0
-->
### Description

<!--
Please add any detail or context that would be useful to a reviewer.
-->

Ensure that execution options are propagated to compound execution group
when creating a PriorityExecutionGroup.

This solves issues where passing `StopOnFirstError` wasn't stopping the
rest of the commands running.

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [x]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).